+++ src/interfaces/bindings/BindingInterface.ts
@@ -2,6 +2,11 @@ import { ContainerInterface } from '../ContainerInterface'
 
 export interface BindingInterface {
   /**
+   * @var {any} value
+   */
+  value: any
+
+  /**
    * Resolve and return the value of the binding.
    *
    * @param  {ContainerInterface} container