+++ src/interfaces/bindings/ResolverBindingInterface.ts
@@ -2,17 +2,7 @@ import { BindingInterface } from './BindingInterface'
 
 export interface ResolverBindingInterface extends BindingInterface {
   /**
-   * Determine if the binding has been resolved at least once.
-   *
-   * @return {boolean}
+   * @var {boolean} hasResolved
    */
-  hasResolved (): boolean
-
-  /**
-   * Set the value of the hasResolved property.
-   *
-   * @param  {boolean} hasResolved
-   * @return {this}
-   */
-  setHasResolved (hasResolved: boolean): this
+  hasResolved: boolean
 }