+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `getValue()` method from `Binding` class
 - `setValue()` method from `Binding` class
 - `getResolver()` method from `ResolverBinding` class
+- `hasResolved()` method from `ResolverBinding` class
+- `setHasResolved()` method from `ResolverBinding` class
 
 ## [0.4.0]
 ### Added